requirements.txt
Flask==2.1.1
googlemaps==4.5.3
pandas==1.3.3
seaborn==0.11.2
matplotlib==3.4.3
folium==0.12.1
celery==5.2.3
redis==4.0.2
python-dotenv==0.19.2
